Rename data directory from `json` to `data`
