+++ scripts/build.sh
@@ -1,6 +1,6 @@
 # Copy input data
-mkdir -p /app/src/json
-cp -rv /mnt/dashboard_input_data/. /app/src/json
+mkdir -p /app/src/data
+cp -rv /mnt/dashboard_input_data/. /app/src/data
 #
 # Build the application
 npm run build
