+++ src/routes/sector_view.svelte
@@ -1,10 +1,10 @@
 <!-- src/routes/sector_view.svelte -->
 <script>
 	import { onMount } from 'svelte';
-	import exposure_stats_data from '../json/data_exposure_stats.json';
-	import techmix_data from '../json/data_techmix_sector.json';
-	import traj_data from '../json/data_trajectory_alignment.json';
-	import emissions_data from '../json/data_emissions.json';
+	import exposure_stats_data from '../data/data_exposure_stats.json';
+	import techmix_data from '../data/data_techmix_sector.json';
+	import traj_data from '../data/data_trajectory_alignment.json';
+	import emissions_data from '../data/data_emissions.json';
 	import { ExposureStatsTile } from '../js/exposure_stats.js';
 	import { techexposure_future } from '../js/techexposure_future.js';
 	import { techmix_sector } from '../js/techmix_sector.js';
